+++ src/main/java/nextstep/auth/authorization/AuthenticationPrincipalArgumentResolver.java
@@ -2,7 +2,7 @@ package nextstep.auth.authorization;
 
 import nextstep.auth.context.Authentication;
 import nextstep.auth.context.SecurityContextHolder;
-import nextstep.member.domain.LoginMember;
+import nextstep.auth.member.LoginMember;
 import org.springframework.core.MethodParameter;
 import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
