+++ src/main/java/nextstep/member/application/LoginMemberService.java
@@ -1,6 +1,6 @@
 package nextstep.member.application;
 
-import nextstep.auth.member.User;
+import nextstep.auth.member.LoginMember;
 import nextstep.auth.member.UserDetailService;
 import nextstep.auth.member.UserDetails;
 import nextstep.member.domain.Member;
@@ -17,7 +17,7 @@ public class LoginMemberService implements UserDetailService {
 
     public UserDetails loadUserByUsername(String email) {
         Member member = memberRepository.findByEmail(email).orElseThrow(RuntimeException::new);
-        return User.of(member.getEmail(), member.getPassword(), member.getRoles());
+        return LoginMember.of(member.getEmail(), member.getPassword(), member.getRoles());
     }
 
     public boolean isUserExist(String email) {
